Refactor: clean up workbook detail
