--- src/app/(after-login)/(top)/workbook-detail/[workbookId]/layout.tsx
@@ -1,7 +1,0 @@
-import React, { ReactNode } from 'react';
-
-const Layout = ({ children }: { children: ReactNode }) => {
-  return <>{children}</>;
-};
-
-export default Layout;
